Ensure node worker is used only in desktop bundle of ipynb extension
