+++ extensions/ipynb/src/ipynbMain.node.ts
@@ -5,9 +5,10 @@
 
 import * as vscode from 'vscode';
 import * as main from './ipynbMain';
+import { NotebookSerializer } from './notebookSerializer.node';
 
 export function activate(context: vscode.ExtensionContext) {
-	return main.activate(context, false);
+	return main.activate(context, new NotebookSerializer(context));
 }
 
 export function deactivate() {
