+++ extensions/ipynb/src/notebookSerializer.web.ts
@@ -3,14 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as vscode from 'vscode';
-import * as main from './ipynbMain';
-import { NotebookSerializer } from './notebookSerializer.web';
+import { NotebookSerializerBase } from './notebookSerializer';
 
-export function activate(context: vscode.ExtensionContext) {
-	return main.activate(context, new NotebookSerializer(context));
-}
-
-export function deactivate() {
-	return main.deactivate();
+export class NotebookSerializer extends NotebookSerializerBase {
 }
